Move functions out of components
