+++ src/utils/post.model.ts
@@ -16,3 +16,7 @@ export interface Tag {
   id: string,
   name: string
 }
+
+export const EmptyPost = (currentUser: BlogUser, createdAt: Timestamp): Post  => {
+  return <Post>{ id: "", title: '', content: '', createdAt: createdAt, updatedAt: createdAt, createdBy: currentUser, updatedBy: currentUser, tags: <Tag[]>[] }
+}
